Q615-043 Adding debug-stack-trace flag
This flag implies debug flag but its use is different as it disallows
catching of exception to give a complete trace to the user. This is to be
used only by developpers.

* src/gnat/gnat_config.ml
(options): The flag implies setting Why3 Debug.stack_trace flag which is
then used to know if this flag is used. It also implies recording the
backtrace.

* */gnat/*.ml
Trying to add a condition when stack_trace is set to reraise exceptions.
The same is done in several places of Why3 for the same reason (getting a
complete trace).

Change-Id: I1e7dfefcf017aab409e0965a7f5c144ba6e80d98