ITP does not use drivers anymore for printing task
it now uses the module core/Pretty, that is generalized so as
to take ident_printer as arguments.
Notice the very nice use of first-class modules !

TODO: a bug remain when printing ident with space in them
TODO: remove the tables in printer_args

We need to discuss with Andrei about the use of "infix " in
infix identifiers which appears to be a problem for parsing
transformation arguments.
Anyway, we don't understand the specific hacks for "mixfix []"
and "mixfix [<-]" in Pretty.ml. Why not similar hacks for "mixfix [..]"
for example?