Using new forward_result to get new results from why3server.
Changing the structure of timeout_handler which now uses one hashtable
with the prover calls information and one queue which has the editor call
information.
The loop in timeout_handler is now done on the result given by why3server
not on a queue of prover task. This is more efficient and the code should
be the same even when blocking = true.