adding new case Undone for pa status. But it does not solve the pb:
interrupted proofs are still impossible to remove.
The pb is that the session does not store the "dynamic" state of the
running session, so it is not possible to check whether a given proof
attempt is in progress or not. The controller should have its own table
to check whether a proofattemptId has a proof in progress or not.