Q927-025 too long socket filename for why3server
The socket generated for why3server's interaction is based on the name
of the current directory path. So, when the project path is very long
(when using option -t of run-test for example), the path+filename exceed
108 characters and the socket cannot be generated. This commit removes
the directory part of the socket name to make it shorter.
Note that this only happened in debug mode or in gnat_server because,
in normal mode, why3server is launched from gnatprove.

* src/driver/prove_client.ml
(connect_internal): Take only the basename of the filepath.

Change-Id: Ifc3e1563f258c696eaee699a9edcb0ecd6849e1a