QA04-053 Fix theory checksum and theory merging with no shapes
This fix the theory checksum which now only uses the immediate subgoals
of the theory instead of using all subgoals. It also reintroduce code that
was lost during itp merge: after merging of a theory, if the theory
checksums match and we did not find detached goals, then we set all the
proof attempt as non obsolete. This is correct because the checksums of
theories match.

* src/session/session_itp.ml
(compute): Computation of theory now takes only immediate subgoals.

(merge_theory): If checksum of theory match we set the whole new theory as
not obsolete.

Change-Id: Iba6bda24fa8bb2e8050931c593643a365323a0b0