QA16-006 Prover called with no limits
When a goal is recognized as valid and obsolete, a prover could be run
without step limits (0). It can actually happen that we change slightly
the generation and provers do not manage to prove the goal anymore. So,
this commit removes the part about obsolete goals in setting the limit.

* src/session/controller_itp.ml
(adapt_limits): Removing the part about obsolete proof.

Change-Id: Ie52c0f1b14b6c729e0b4a2c71d7a4586d2758dac