Make theories/int.why depend on itself when generating its documentation.
When theories/int.why is handled, it is missing from the environment, so
Why3 cannot find integer operators to handle some constructs, e.g.
variants. So this commit duplicates the file before generating the
documentation. Luckily, links are nevertheless correct.

This hack would not be needed if integer operators were builtin.