Update configure and regtests / nightly-build so that realization are generated again.
* Makefile.in
Realizations for isabelle are generated even if Isabelle is not installed.
The .thy file are checked only if isabelle is installed. This allows us
to treat it as with coq and be able to detect changes in realizations
with a script.
Dont erase xml generated file during clean.

* configure.in
Changed the configure to always generate a default isabelle version (set
to 2017) for generation of the realization driver (needed even when no
isablle support is given).

* examples/regtests.sh
Adding option --check-realization and --only-realization which add a test
the compares the output of the new realization compared to the one
recorded. It complains if they are different forcing the user to update the
realization files.

* lib/isabelle/*
Adding xml files corresponding to generated realization for Isabelle. This
allows for the abovementionned script to work even if Isabelle is not
installed (which is the case for most people editing theories).

* examples/nightly-bench.sh
Add the option for realization check

* misc/ci-bench
Calls the test for realization

Conflicts:
	examples/regtests.sh