Fixes #59
Adding random (to be set) shortcut to the functions: next unproven goal,
expand, expand_all, collapse, parent, first_child. These can be accessed
from the command line by typing the command (list_ide_command is for help
on these commands).
Fixing shortcuts with ~key instead of add_accelerator.
[get_first_unproven_goal_around cn] now returns a different node from cn
and, if a brother of cn is returned, it returns preferably one placed
after cn in the tree.