Create directories bin and lib/plugins on the fly (bug #21566).
Some people dump the content of tarballs into git repositories and thus
experience compilation failures. Indeed, some directories from the tarball
are empty and thus not preserved by git. So this commit creates these
directories on the fly.

As for plugins/printer, it has to be handled differently, since it is
needed by ocamldep and thus cannot be created on the fly. So its .keepme
file is put in the tarball to keep it nonempty.