R205-027 Changing timeout of gnat_server
This commit adds a debug flag for gnat_server and patch several timing
problems in the server. The timeout for select is now always positive and
the maximum delay allowed is 0.2 seconds.

* src/gnat/gnat_config.ml
Adding debug flag --debug-gnat-server

* src/gnat/gnat_server.ml
(debug_to_file): print debug info to a file when --debug-gnat-server is
given.
(main_loop): Changed the definition of timeout so that it is never
negative. Add debugging information as calls to debug_to_file.

Change-Id: I34ce7baf8830631f7e25f300cd0fb603e8313428