Ity: in a private type, modification of ghost fields is visible
This is important for the "loose" fields: mutable components of
a private type which can be modified without a direct write into
the containing object. These subcomponents can be modified by
abstract functions, and these modifications are actually visible,
since a refinement of a private type can make ghost fields non-ghost.