Add debug keep_vcs: allow to use a given filename to save prover files.
It is now possible to give an optional argument to schedule_proof_attempt
so that (when in debug mode) the file given is used to store the generated
input file of the prover. In non debug mode, the file given to
schedule_proof_attempt is used but removed after the call to the prover.

Currently, no file is ever given to schedule_proof_attempt but this can be
used by people using Why3 as a backend.