Tentatively add new error handling mode of menhir
This commit tries to add the new parsing errors from menhir. It changes
the lexer so that precise error made from the provided handcrafted.messages
can be displayed on failure.
Also fix #111

* Makefile.in
This adds a rule for src/parser/parser_messages.ml which either compile
the messages or regenerates them. It needs regeneration when changes are
added in .mly or .messages file. This also adds option .DELETE_ON_ERROR.

* src/parser/lexer.mll
The parser now uses loop_handle function from menhir (see menhir ref.
manual).

* src/parser/report.ml
Added file from Compcert's preparser to report error from "checkpoint". It
allows to use $0 notation in handcrafted.messages

* src/parser/handcrafted.messages
To be edited file for error messages.

* bench/*
Adding a parsing-bench for parsing errors following the model of ce-bench:
an oracle is provided so that error messages are not lost.