R504-027 Merg of why3
This allows the merge commit 99f9d000 to compile.

* drivers/*.drv
New drivers are necessary to use the counterexamples. Instead of
being added to task by Ocaml code, the cntexample meta is now added inside
driver's code. This meta was added to appropriate drv files.

* src/gnat/gnat_objectives.ml
reload_files/add_file now raise error instead of returning them. The
unnecessary cntexample arguments are also removed.

* src/gnat/gnat_counterexamples.ml
Parsing of let expression in counterexample model imposes that arrays can
be a single variable. This should be substituted in collect_data_model and
should not appear in legal data in gnat_counterexamples.ml (still we need
to state that the case can happen).

Change-Id: If5efdbaefaab2ca3145bc9011e0a7db71182eb30