Handle signal information in the client.
The server was properly detecting that a prover had been killed by a
signal, but the client was ignoring this information and was just assuming
that the client had exited normally. Now the client properly marks the
prover as killed by a signal. Also, the server now transmits the signal
number on Linux (that was already the case on Windows).