Fuzzy a bit more the detection of timeouts by 0.1 second.
The formula now matches the one for detecting variations on Unknown
answers. This commit also restores the time limit for Metis, since its
removal simply caused Metis to be killed just as early by the operating
system.

Theoretically, all the provers suffer from this issue, so it is kind of
a mystery why Metis is the only prover to be killed so early. I guess
the value of ru_utime is not reliable for Metis.