Check whether ocamlfind might have returned the actual location of nums.cma.
The standard install location of the num library is $OCAMLLIB. Thus, the
answer of ocamlfind cannot generally be trusted since it returns the
location of the META file. Yet, on Arch Linux, nums.cma is actually stored
next to its META file, for some reason. So, this commit makes configure
try this location before falling back to the standard one.