Q313-019 Disable interp_float
interp_float is used in Why3 to change the representation of float number
into something involving hexadecimal. This disables it so that the float
representation are simpler.

* src/core/model_parser.m*
Copy of the corresponding patch in Why3 (add optional argument to
interp_float)

* src/driver/parse_smtv2_model_lexer.mll
Use the optional arguments for interp_float to disable hexadecimal
printing.

Change-Id: Iee36de32a5a739e93e1fe82b7cd69e675198a189