R722-002 Handle the case where a non existent node is sent to Why3
When no node is selected in SPARK manual proof, the IDE sends the node
number 0 which corresponds to the root_node in Why3 (not in model_any map).
This makes Why3 sends a Not_Found exception.
This commit allows the mapping not to be complete in which case we return
a message to the IDE: "Please select a node id". This should work for any
node_id used in IDE that does not exists in Why3.

* src/session/itp_server.ml
(any_from_node_ID): return an option type
(is_below): Adapt type.
(send_task): Ask to select a node id to send a task
(schedule_proof_attempt): Ask to select a node id when none selected
(schedule_edition): Ask to select a node id when none selected
(schedule_transformation): Ask to select a node id when none selected
(run_strategy): Same
(remove_node): Same
(noify_first_unproven_node): Same
(treat_request): Same

Change-Id: I93b032095b4b41b98f104282a1c0078499400d3c