R722-002 Allow selection at startup
When the IDE starts, it does not know on which node it is because of
automatic focusing on interesting nodes (done by the server).
When the IDE asks which node it should go to, if the same node is found by
the server, then the server answers nothing.
With this patch, the server has a mode in which it always answers one 
node when asked which one is the next unproven node.

* src/session/itp_server.ml
(get_first_unproven_goal_around): always_send is set to true.

* src/session/server_utils.ml
(get_first_unproven_goal_around): Add argument always_send so that 
it always gives an unproved node as result. For Why3, this is always set 
to false.
Change the None case.

Change-Id: Ieebf5af5e866ec4cba9af55dfc3a727296eb3e8d