Q426-004 add --limit_region to restrict proof to a range of lines
* src/gnat/gnat_loc.mli
(region): Type for a region (a file and a pair of lines).
(mk_region): Construct a region with given file and first and last line.
(print_region): Printer for debug purposes.
(in_region): Check whether a location is inside a region.

* src/gnat/gnat_config.mli
(limit_region): set if option --limit-region was given; we only prove VCs
 in the region.

* src/gnat/gnat_objectives.ml
(add_to_objective): Filter according to the specified region if any.

Change-Id: Ie017e40e44eeb6b32d8e22f6e31ef06f84058765