R716-031 Adding location of the why3.conf file on driver error
When a user defined driver fails to launch, and the user tries to use Coq
(for example) the location of the why3.conf file that contains the driver
part is now mentioned as part of the error.

* src/gnat/gnat_config.ml
(find_driver_file): Now takes an additional argument: the path to the
configuration file.
(get_gnatprove_config): Add the additional argument to find_driver_file.

Change-Id: I2ff83e29f573f0af1089d8e36e603728faa9e640