RA26-009 Change .mlw at reload
This commit replace the .mlw file registered in the session by the input
file at reload.

* src/gnat/gnat_objectives.ml
(init_cont): We search for the file inside the session. If the file is the
same as the input gnatwhy3 .mlw file, we do nothing. If the file is
another one, we replace the file in the session. If there are several
files in the session we fail: this should not happen.

Change-Id: Ie5eda60373da0a9d9efb89e4cc7a996b4a1309a8