RC04-029 Add parsable error report for gnatwhy3 exception
When uncaught exception such as assert false occurs, the output seen in
gnatprove should be more clear because abort_with_message is used.
It also adds backtrace generation for stack_trace of Why3.

* src/gnat/gnat_util.ml
(): Register a function to print uncaught exception which should be used
when assert false is returned for example.

* src/gnat/gnat_config.ml
(options): Add recording of backtrace for stack_trace flag (not only for
--debug-stack-trace flag)

Change-Id: Idc98f53c25a844ec67aba97fa591572869f3a386