S110-004 Chery-pick of 06c1a5a5 and 9fb35c19 from why3 master
Reasons for cherry pick:
- 06c1a5a5: use file_path so that we don't get errors when querying from
files of sessions (between Windows and Linux)
- 9fb35c19: side effect of the above to have a recent version of
Sysutil.relativize_filename.

This should solve the problem because this makes all the saved filename
printed with "/" even on Windows. Filename can be loaded with either
forward slash or backslash but always get printed with "/".

There may still be some problems with the way things are printed for Coq
or Isabelle proof.

Change-Id: Id8efc8cbc190709a428df47bafbdd4f981d7442a