ce: Add a special case for projected map inside record with one field
Record with one field have a specific treatment in VC generation. The
combination of this with the projections of first and last elements of
arrays (that are inside records with one field) were not considered. Hence,
the necessity of this hackish solution which treat specific case of
First and Last.