Restore drivers.
This restore part of the drivers from gnatwhy3 before newsystem. The
drivers were overridden by those in Why3 master. This is now in discussion
whether Why3 master can add some of this patch in its master. In
particular, the new syntax function could help provers.

This was originally written by Claire when asserting the diffs of
newsystem.

Change-Id: I67d8ed2e49d8aec5f91505c251a86eba403a4858