Adhoc modification of eliminate_epsilon
The modification make sure that realizations will be queried the same as in
Why3 both for Isabelle and Coq.
Also, the former optimization that consists in removing this small part
of the code altogether should still work most of the time (as SPARK does
not generate lambda expression and we need to have lambda expression to
go in this part of the code).

Change-Id: I1a62bfd267f7b8539e056f91679c06bebe7ecc78