S313-020 Fix cvc4_qf
This fixes the transformation eliminate_unknown_types so that it properly
remove anything that depends on given types. Previously, some constant or
definitions containing constants of removed types could survive which make
the transformation fail.

* src/transform/eliminate_unknown_types.ml
Use Trans.bind to avoid having transformations defined inside
Trans.on_metas which is a cause of error.
Use of Debug.dprintf.
(remove_terms): Add already_removed which keeps track of already removed
lsymbols. When such lsymbols is found in a definition, the declaration is
also removed.

Change-Id: I5de3dd368d87f18c0bb12cad0e92f9f3b0fe7d28