S314-023 Out_of_memory is not an internal error
There are two ways an ocaml application can behave when running out
of memory.
1) An exception Out_of_memory is raised, which can be caught.
2) The ocaml programs stops and prints a message "Fatal error: out
   of memory" on stderr.

This commit is only about (1). Previously, any exception raised in
gnatwhy3 would caught and be reported as an "internal" error that is
printed with a bugbox. We now special-case Out_of_memory to set
"internal" to false, so no bugbox is printed (but the error message
is still reported to the user).

This special-casing is done in the uncaught_exception_handler. To
reach this handler in the case of Out_of_memory, we add a case
before any catch-all handlers in gnatwhy3 to reraise that exception.

* gnat_main.ml
(main) reraise Out_of_Memory
* gnat_objectives.ml
(init_cont): reraise Out_of_Memory
* gnat_util.ml
(uncaught_exception_handler): if the exception was Out_of_Memory,
  set the "internal" flag to false

Change-Id: I137f6271e154c0947690168b7077213adeaeb997