S325-010 Remove z3-realize.drv
This file is used nowhere. The only reference found was for a Why3 (from
Why3 master) tests to check a kind of smt realizability: seems to be
asking CVC4/Z3 to prove the theories they are given.

This was removed in f05478680f8ecee in Why3 master. It is not used in
SPARK and was probably kept because of a merge mistake.

* drivers/z3-realize.drv
Remove file

Change-Id: Ia40d0c40f124d1cbbb7a518f33ba0ac922b49c27