Change the internal representation of shapes and shape files
This commit changes the way shapes are represented. They are no longer
individual string of 1024 characters for each task.

The idea is that shapes are repeated through different task. So, we now
have a mapping int -> shapes which contains shapes of declarations. This
mapping is in the session. The shapes are now given as a list of
successive integer of this mapping.
The shape files are saved in the same way:
- first, the shapes of all declarations that were converted to shape
- second, checksum followed by a string representing the succession of
variables used in a specific shape.

This is a first commit on change for shapes. This should not change the
way pairing is done. For pairing, only the two first variables are used:
explanation and goal.
In the future, we plan to use the other local declarations as well.

Related changes:
- Transform to shape only the local declarations
- The shape size limit is now per declarations

Deprecation:
- Termcode.equal_shape is removed
- Session_itp.empty_session now requires a shape_version