S416-010 Treat is_trivial as a transformation
In the former state, some goals that are tagged as trivial could be saved
in the session as unproved. When replaying, these goals could be confused
with other trivial goals (not at toplevel) which were not tagged as
trivial and got a prover. This confusion lead to bad behavior of replay.

This commit changes is_trivial into a transformation so that the
information is kept in the session. The goal that is trivial will be the
parent of the "trivial_true" transformation. The complexity should be
exactly the same as is_trivial was called on all top level goals.

Expected changes: a lot of goals will appear to be proved by Trivial
instead of CVC4 in the test.out.

* Makefile.in
Add gnat_trivial in the source

* src/gnat/gnat_main.ml
(register_goal): is_trivial is changed to: apply the trivial transformation
and after look if the transformation is in the corresponding goal

* src/gnat/gnat_objectives.ml
(apply_trivial): Add apply_trivial which just applies trivial_true on a
goal.

* src/gnat/gnat_trivial.ml
Copy the is_trivial function from gnat_main here.

Change-Id: Ie939772faaf0ec2c2896d4ff9737431c7527cee8