S509-002 Remove socket in debug mode or gnat_server mode
When in debug mode or gnat_server mode, the socket for why3server is
created internally by Why3. In this case, we fail to clean the socket.
This makes the unsafe copy of the gnatprove directory (done by GPS) to
raise errors. This patch ensures that on correct execution, the socket is
removed (when created internally).

* src/driver/prove_client.ml
(socket_path): Global variable holding the socket_path
(remove_socket): New function which removes the socket if the file exists.
(client_disconnect): Now call remove_socket.

Change-Id: Idf80ab4c4e16a1189ad157dd0b86399a5364183d