Extending pairing
Before, the pairing was based on the least common prefix of the
concatenation of explanation and shape of goals. When those are not enough
to determine how pairing should be done, we add a light comparison on the
rest of the hypothesis: we compare the integers list (more common
hypothesis mean a better chance to be paired).